ESP Processing
Backup of file when trying to process data. Do not need to do this but just in case

+++ UGV-ESP/UGV-ESP-V2/network_defines.hpp
@@ -52,7 +52,7 @@ typedef struct
     double y;         // y position
     double v;         // Linear velocity
     double theta;     // Heading
-    uint64_t ts_ms;   // Timestamp in ms
+    uint32_t ts_ms;   // Timestamp in ms
     node_state state; // Current executing state of the node
     double x_exp;     // expected x position relative to the start based on path
     double y_exp;     // expected y position relative to the start based on path
@@ -70,13 +70,13 @@ typedef struct
     double y;       // y position
     double v;       // Linear velocity
     double theta;   // Heading
-    uint64_t ts_ms; // Timestamp in ms
+    uint32_t ts_ms; // Timestamp in ms
 } packet_path_point;
 
 
 typedef struct 
 {
-    double ts_ms; // Timestamp in ms
+    uint32_t ts_ms; // Timestamp in ms
     double v_right; // current velocity of the right motor
     double d_right; // cumulative distance the right motor has traveled since time 0
     double v_left; // current velocity of the left motor
